backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
firebase-admin
python-multipart
PyPDF2